Move Last condition
