+++ Common/src/main/java/corgitaco/mobifier/common/MobifierConfig.java
@@ -11,7 +11,6 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import corgitaco.corgilib.entity.condition.BiomeTagCondition;
 import corgitaco.corgilib.entity.condition.InDimensionCondition;
 import corgitaco.mobifier.Mobifier;
-import corgitaco.mobifier.common.condition.Condition;
 import corgitaco.mobifier.common.util.DoubleModifier;
 import corgitaco.mobifier.common.util.MobifierUtil;
 import corgitaco.mobifier.mixin.access.AttributeSupplierAccess;
@@ -46,8 +45,8 @@ public record MobifierConfig(boolean dumpRegistries, Map<EntityType<?>, List<Mob
             }
 
         }), true, new ArrayList<>(), Util.make(new ArrayList<>(), (list1) -> {
-            list1.add((Condition) new BiomeTagCondition(ImmutableList.of(BiomeTags.HAS_DESERT_PYRAMID)).codec());
-            list1.add((Condition) new InDimensionCondition(ImmutableList.of(Level.OVERWORLD)).codec());
+            list1.add(new BiomeTagCondition(ImmutableList.of(BiomeTags.HAS_DESERT_PYRAMID)));
+            list1.add(new InDimensionCondition(ImmutableList.of(Level.OVERWORLD)));
         })))));
         Registry.ENTITY_TYPE.forEach(entityType -> {
             if (entityType != EntityType.HUSK) {
